WIP 6 (Semifinal WIP Build?)
The RDB still needs more work and all the games listed in it must be tested (does TLB need to be set off for some of these games? TLB being disabled was causing crashes in tandem with Limit FPS disabled) but it's looking pretty solid now. Next WIP build may be the last and then the build after would be the first build out of WIP stage, maybe call that "Release 1" or something.