see description
Alright so this puts AI audio on by default which is what I want, and not able to be set off. As there is no reason for it to be off and it seems to have multiple benefits from having it on, it should not have ever been toggle-able. I have also added LLE Slow mode, with LLE plugins, interpreter, and counter factor 3 for more general stability as an experimental option. However, I think now with abl removed, we can see how that was setup as a checkbox option and convert ABL from a dropdown option, to a checkbox option, and then remove off/on identifiers and replace the positioning of the menudes and warn for SLOW with where off/on is in the language files. so we will get to trying that next and then see what else I can think of to improve project64.